validation first, getting auth_info after that. Also move authentication model to model directory
