+++ management_api_app/tests_ma/test_api/conftest.py
@@ -17,14 +17,14 @@ def no_database():
 
 
 def override_get_user():
-    from services.authentication import User
+    from models.domain.authentication import User
     return User(id="1234", name="test", email="test", roles=[""], roleAssignments={"ab123": "ab124"})
 
 
 @pytest.fixture(scope='module')
 def admin_user():
     def inner():
-        from services.authentication import User
+        from models.domain.authentication import User
         return User(id="1234", name="test", email="test", roles=["TREAdmin"], roleAssignments={"ab123": "ab124"})
     return inner
