+++ management_api_app/tests_ma/test_api/test_routes/test_workspaces.py
@@ -10,7 +10,7 @@ from models.domain.resource import Status, Deployment
 from models.domain.workspace import Workspace
 from resources import strings
 from api.routes.workspaces import get_current_user
-from services.authentication import User
+from models.domain.authentication import User
 
 
 pytestmark = pytest.mark.asyncio
@@ -100,8 +100,7 @@ async def test_workspaces_id_get_returns_workspace_if_found(get_workspace_mock, 
 @ patch("api.routes.workspaces.send_resource_request_message")
 @ patch("api.routes.workspaces.WorkspaceRepository.save_workspace")
 @ patch("api.routes.workspaces.WorkspaceRepository.create_workspace_item")
-@ patch("api.routes.workspaces.extract_auth_information", return_value={})
-async def test_workspaces_post_creates_workspace(extract_auth_info_mock, create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
+async def test_workspaces_post_creates_workspace(create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
     app.dependency_overrides[get_current_user] = admin_user
     workspace_id = "000000d3-82da-4bfc-b6e9-9a7853ef753e"
     create_workspace_item_mock.return_value = create_sample_workspace_object(workspace_id)
@@ -117,8 +116,7 @@ async def test_workspaces_post_creates_workspace(extract_auth_info_mock, create_
 @ patch("api.routes.workspaces.WorkspaceRepository.save_workspace")
 @ patch("api.routes.workspaces.WorkspaceRepository.create_workspace_item")
 @ patch("api.routes.workspaces.WorkspaceRepository._validate_workspace_parameters")
-@ patch("api.routes.workspaces.extract_auth_information", return_value={})
-async def test_workspaces_post_calls_db_and_service_bus(extract_auth_info_mock, validate_workspace_parameters_mock, create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
+async def test_workspaces_post_calls_db_and_service_bus(validate_workspace_parameters_mock, create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
     app.dependency_overrides[get_current_user] = admin_user
     workspace_id = "000000d3-82da-4bfc-b6e9-9a7853ef753e"
     validate_workspace_parameters_mock.return_value = None
@@ -135,8 +133,7 @@ async def test_workspaces_post_calls_db_and_service_bus(extract_auth_info_mock, 
 @ patch("api.routes.workspaces.WorkspaceRepository.save_workspace")
 @ patch("api.routes.workspaces.WorkspaceRepository.create_workspace_item")
 @ patch("api.routes.workspaces.WorkspaceRepository._validate_workspace_parameters")
-@ patch("api.routes.workspaces.extract_auth_information", return_value={})
-async def test_workspaces_post_returns_202_on_successful_create(extract_auth_info_mock, validate_workspace_parameters_mock, create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
+async def test_workspaces_post_returns_202_on_successful_create(validate_workspace_parameters_mock, create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
     app.dependency_overrides[get_current_user] = admin_user
     workspace_id = "000000d3-82da-4bfc-b6e9-9a7853ef753e"
     validate_workspace_parameters_mock.return_value = None
@@ -153,8 +150,7 @@ async def test_workspaces_post_returns_202_on_successful_create(extract_auth_inf
 @ patch("api.routes.workspaces.WorkspaceRepository.save_workspace")
 @ patch("api.routes.workspaces.WorkspaceRepository.create_workspace_item")
 @ patch("api.routes.workspaces.WorkspaceRepository._validate_workspace_parameters")
-@ patch("api.routes.workspaces.extract_auth_information", return_value={})
-async def test_workspaces_post_returns_503_if_service_bus_call_fails(extract_auth_info_mock, validate_workspace_parameters_mock, create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
+async def test_workspaces_post_returns_503_if_service_bus_call_fails(validate_workspace_parameters_mock, create_workspace_item_mock, save_workspace_mock, send_resource_request_message_mock, app: FastAPI, client: AsyncClient, admin_user: User):
     app.dependency_overrides[get_current_user] = admin_user
     workspace_id = "000000d3-82da-4bfc-b6e9-9a7853ef753e"
     validate_workspace_parameters_mock.return_value = None
@@ -170,8 +166,7 @@ async def test_workspaces_post_returns_503_if_service_bus_call_fails(extract_aut
 
 @ patch("api.routes.workspaces.WorkspaceRepository._get_current_workspace_template")
 @ patch("api.routes.workspaces.WorkspaceRepository._validate_workspace_parameters")
-@ patch("api.routes.workspaces.extract_auth_information", return_value={})
-async def test_workspaces_post_returns_400_if_template_does_not_exist(extract_auth_info_mock, validate_workspace_parameters_mock, get_current_workspace_template_mock, app: FastAPI, client: AsyncClient, admin_user: User):
+async def test_workspaces_post_returns_400_if_template_does_not_exist(validate_workspace_parameters_mock, get_current_workspace_template_mock, app: FastAPI, client: AsyncClient, admin_user: User):
     app.dependency_overrides[get_current_user] = admin_user
     validate_workspace_parameters_mock.return_value = None
     get_current_workspace_template_mock.side_effect = EntityDoesNotExist
